admin: handle messages consisting of spaces
As string:tokens returns an empty array when tokenizing a string of one
or more spaces with space as a separator, the assignment resulted in a
bad match if anyone sent one or more spaces (and nothing else) in a
message, causing a crash. This commit handles this edge case.